services/ml_engine/requirements.txt
kafka-python>=2.0.2
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
colorlog>=6.7.0
joblib>=1.3.0
